requirements.txt
azure-storage-blob>=12.22.0
azure-identity>=1.19.0
openai>=1.59.0
python-dotenv>=1.0.1
